Allow Customization of the Element Header (#21)
* Allow Customization of the Element Header

HeaderClass expands on the HeaderStyle attribute of an element to allow you to define a custom class to be applied to the headers as well as changing the style.

You can now add a Header tag to either the Element or Grid that allows you to insert custom markup that is used instead of the default header markup. When this markup is used the Header level and alignment are ignored. This means the header does not just have to contain a title but can have varying test, images icons etc.

* Removed the gitignore file.

* Add example and sample for setting custom headers in a grid and dynamically opening edit mode.

* Update custom header example with edit mode that allows move and resize of the elements.